::cmd is DEAD, long live ::command!
