+++ src/command/logs.rs
@@ -9,7 +9,7 @@ use rustyline::completion::Pair as RustlinePair;
 use strfmt::strfmt;
 
 use crate::{
-    cmd::{exec_match, start_clap, Cmd},
+    command::command_def::{exec_match, start_clap, Cmd},
     command::{parse_duration, valid_date, valid_duration, valid_u32},
     completer,
     env::Env,
