+++ src/command/namespaces.rs
@@ -4,8 +4,8 @@ use k8s_openapi::api::core::v1 as api;
 use rustyline::completion::Pair as RustlinePair;
 
 use crate::{
-    cmd::{exec_match, start_clap, Cmd},
-    command::{run_list_command, sort_arg, Extractor},
+    command::command_def::{exec_match, sort_arg, start_clap, Cmd},
+    command::{run_list_command, Extractor},
     completer,
     env::Env,
     kobj::{KObj, ObjType},
