+++ src/command/pods.rs
@@ -5,11 +5,10 @@ use ansi_term::{
 use clap::{App, Arg};
 use k8s_openapi::api::core::v1 as api;
 use k8s_openapi::ListOptional;
-use rustyline::completion::Pair as RustlinePair;
 
 use crate::{
-    cmd::{exec_match, start_clap, Cmd},
-    command::{run_list_command, show_arg, sort_arg, Extractor},
+    command::command_def::{exec_match, show_arg, sort_arg, start_clap, Cmd},
+    command::{run_list_command, Extractor},
     completer,
     env::{Env, ObjectSelection},
     kobj::{KObj, ObjType},
