+++ src/command/replicasets.rs
@@ -1,11 +1,10 @@
 use ansi_term::Colour::Yellow;
 use clap::{App, Arg};
 use k8s_openapi::api::apps::v1 as apps_api;
-use rustyline::completion::Pair as RustlinePair;
 
 use crate::{
-    cmd::{exec_match, start_clap, Cmd},
-    command::{keyval_string, run_list_command, show_arg, sort_arg, Extractor},
+    command::command_def::{exec_match, show_arg, sort_arg, start_clap, Cmd},
+    command::{keyval_string, run_list_command, Extractor},
     completer,
     env::Env,
     kobj::{KObj, ObjType},
