+++ src/main.rs
@@ -61,8 +61,6 @@ extern crate url;
 extern crate yasna;
 
 #[macro_use]
-mod cmd;
-#[macro_use]
 mod command;
 #[macro_use]
 mod completer;
